+++ lib/arachni/support/cache/preference.rb
@@ -9,33 +9,20 @@
 module Arachni
 module Support::Cache
 
-# Random Replacement cache implementation.
-#
-# Better suited for low-latency operations, discards entries at random
-# in order to make room for new ones.
-#
 # @author Tasos "Zapotek" Laskos <tasos.laskos@arachni-scanner.com>
 class Preference < Base
-
-    # Storage method
-    #
-    # @param    [Object]    k
-    #   Entry key.
-    # @param    [Object]    v
-    #   Object to store.
-    #
-    # @return   [Object]
-    #   `v`
-    def store( k, v )
-        prune if capped? && (size > max_size - 1)
-        cache[make_key( k )] = v
-    end
 
     def prefer( &block )
         @preference = block
     end
 
     private
+
+    def store_with_internal_key( k, v )
+        prune if capped? && (size > max_size - 1)
+
+        @cache[k] = v
+    end
 
     def find_preference
         @preference.call